requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
beautifulsoup4==4.12.2
python-dotenv==1.0.0
mutagen==1.46.0
